requirements.txt
flask
werkzeug
python-dotenv
authlib
requests
